+++ ha_addon/run.sh
@@ -14,6 +14,7 @@ if [ -f "$CONFIG_PATH" ]; then
     export BOOST_SPEED=$(jq --raw-output '.boost_speed' $CONFIG_PATH)
     export NORMAL_SPEED=$(jq --raw-output '.normal_speed' $CONFIG_PATH)
     export BOOST_DURATION_MINUTES=$(jq --raw-output '.boost_duration_minutes' $CONFIG_PATH)
+    export HUMIDITY_VERY_HIGH_THRESHOLD=$(jq --raw-output '.humidity_very_high_threshold' $CONFIG_PATH)
     export HUMIDITY_HIGH_THRESHOLD=$(jq --raw-output '.humidity_high_threshold' $CONFIG_PATH)
     export HUMIDITY_LOW_THRESHOLD=$(jq --raw-output '.humidity_low_threshold' $CONFIG_PATH)
     export NIGHT_START=$(jq --raw-output '.night_start' $CONFIG_PATH)